+++ spec/rack_attack_throttle_spec.rb
@@ -17,7 +17,7 @@ describe 'Rack::Attack.throttle' do
     before { get '/', {}, 'REMOTE_ADDR' => '1.2.3.4' }
 
     it 'should set the counter for one request' do
-      key = "rack::attack:#{Time.now.to_i / @period}:ip/sec:1.2.3.4"
+      key, _ = Rack::Attack.cache.send(:key_and_expiry, "ip/sec:1.2.3.4", @period, true)
       _(Rack::Attack.cache.store.read(key)).must_equal 1
     end
 
@@ -27,6 +27,7 @@ describe 'Rack::Attack.throttle' do
         limit: 1,
         period: @period,
         epoch_time: Rack::Attack.cache.last_epoch_time.to_i,
+        retry_after: Rack::Attack.cache.last_retry_after_time.to_i,
         discriminator: "1.2.3.4"
       }
 
@@ -52,6 +53,7 @@ describe 'Rack::Attack.throttle' do
         limit: 1,
         period: @period,
         epoch_time: Rack::Attack.cache.last_epoch_time.to_i,
+        retry_after: Rack::Attack.cache.last_retry_after_time.to_i,
         discriminator: "1.2.3.4"
       )
 
@@ -73,7 +75,7 @@ describe 'Rack::Attack.throttle with limit as proc' do
     before { get '/', {}, 'REMOTE_ADDR' => '1.2.3.4' }
 
     it 'should set the counter for one request' do
-      key = "rack::attack:#{Time.now.to_i / @period}:ip/sec:1.2.3.4"
+      key, _ = Rack::Attack.cache.send(:key_and_expiry, "ip/sec:1.2.3.4", @period, true)
       _(Rack::Attack.cache.store.read(key)).must_equal 1
     end
 
@@ -83,6 +85,7 @@ describe 'Rack::Attack.throttle with limit as proc' do
         limit: 1,
         period: @period,
         epoch_time: Rack::Attack.cache.last_epoch_time.to_i,
+        retry_after: Rack::Attack.cache.last_retry_after_time.to_i,
         discriminator: "1.2.3.4"
       }
 
@@ -104,7 +107,7 @@ describe 'Rack::Attack.throttle with period as proc' do
     before { get '/', {}, 'REMOTE_ADDR' => '1.2.3.4' }
 
     it 'should set the counter for one request' do
-      key = "rack::attack:#{Time.now.to_i / @period}:ip/sec:1.2.3.4"
+      key, _ = Rack::Attack.cache.send(:key_and_expiry, "ip/sec:1.2.3.4", @period, true)
       _(Rack::Attack.cache.store.read(key)).must_equal 1
     end
 
@@ -114,6 +117,7 @@ describe 'Rack::Attack.throttle with period as proc' do
         limit: 1,
         period: @period,
         epoch_time: Rack::Attack.cache.last_epoch_time.to_i,
+        retry_after: Rack::Attack.cache.last_retry_after_time.to_i,
         discriminator: "1.2.3.4"
       }
 
@@ -135,7 +139,7 @@ describe 'Rack::Attack.throttle with block retuning nil' do
     before { get '/', {}, 'REMOTE_ADDR' => '1.2.3.4' }
 
     it 'should not set the counter' do
-      key = "rack::attack:#{Time.now.to_i / @period}:ip/sec:1.2.3.4"
+      key, _ = Rack::Attack.cache.send(:key_and_expiry, "ip/sec:1.2.3.4", @period, true)
       assert_nil Rack::Attack.cache.store.read(key)
     end
 
@@ -163,7 +167,7 @@ describe 'Rack::Attack.throttle with throttle_discriminator_normalizer' do
 
   it 'should not differentiate requests when throttle_discriminator_normalizer is enabled' do
     post_logins
-    key = "rack::attack:#{Time.now.to_i / @period}:logins/email:person@example.com"
+    key, _ = Rack::Attack.cache.send(:key_and_expiry, "logins/email:person@example.com", @period, true)
     _(Rack::Attack.cache.store.read(key)).must_equal 3
   end
 
@@ -174,7 +178,7 @@ describe 'Rack::Attack.throttle with throttle_discriminator_normalizer' do
 
       post_logins
       @emails.each do |email|
-        key = "rack::attack:#{Time.now.to_i / @period}:logins/email:#{email}"
+        key, _ = Rack::Attack.cache.send(:key_and_expiry, "logins/email:#{email}", @period, true)
         _(Rack::Attack.cache.store.read(key)).must_equal 1
       end
     ensure